Move awstesting.SortedKeys to private/util package (#721)
This dependency caused the main packages to import awstesting which in
turn brought a lot of testing related dependencies with itself. By
moving a single file we can get rid of this dependency.
