+++ private/util/sort_keys.go
@@ -1,4 +1,4 @@
-package awstesting
+package util
 
 import "sort"
